requirements.txt
anyio==4.6.2.post1
asgiref==3.8.1
Brotli==1.1.0
certifi==2024.8.30
clients==1.5
distlib==0.3.9
dj-database-url==2.3.0
Django==5.1.3
djangorestframework==3.15.2
filelock==3.16.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
mysqlclient==2.2.6
packaging==24.2
platformdirs==4.3.6
psycopg2-binary==2.9.10
PyMySQL==1.1.1
sniffio==1.3.1
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
virtualenv==20.27.1
whitenoise==6.8.2
